requirements.txt
django==3.2
django-environ==0.11.2
djangorestframework==3.15.2